requirements.txt
aiohttp
python-dotenv
requests
nltk
sentence-transformers
